add casts to avoid warnings in visual c++ 2005
git-svn-id: svn://svn.code.sf.net/p/tivodecode/code/tivodecode/trunk@222 3cae7af1-9334-450b-b769-427e0023cdf8

+++ tivodecode.c
@@ -290,7 +290,7 @@ int main(int argc, char *argv[])
             }
 
             prepare_frame(&metaturing, 0, 0);
-            skip_turing_data(&metaturing, chunk_start - current_meta_stream_pos);
+            skip_turing_data(&metaturing, (size_t)(chunk_start - current_meta_stream_pos));
             decrypt_buffer(&metaturing, chunk->data, chunk->data_size);
             current_meta_stream_pos = chunk_start + chunk->data_size;
